refactoring epics client into separate files
